Added a packaging structure, refactored code and added comments.
