--- src/main/BioRetrieval.java
@@ -1,3 +1,0 @@
-public interface BioRetrieval {
-    public String getBio();
-}
