+++ src/main/Main.java
@@ -1,3 +1,10 @@
+import Managers.Manageable;
+import Managers.Manager;
+import Managers.PrioritySkillManager;
+import MatchingAlgoHelpers.MatchMaker;
+import MatchingAlgoHelpers.Scheduler;
+import UserData.Player;
+
 import java.util.Scanner;
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -238,8 +245,6 @@ public class Main {
     public static void presentMatches(){
             Scheduler scheduler1 = new Scheduler();
             MatchMaker matchMaker1 = new MatchMaker();
-            //TODO The line below needs to be modified. Manager is now an abstract class.
-            //TODO so we need to either call PrioritySkillManager or PriorityLocationManager.
             Manager managerObject = new PrioritySkillManager(matchMaker1, scheduler1);
             ArrayList<Manageable> users_in_order = new ArrayList<>();
             users_in_order = managerObject.getRankedList(user, all_users);
