--- src/main/Manageable.java
@@ -1,5 +1,0 @@
-import java.util.ArrayList;
-
-public interface Manageable extends Matchable, Schedulable{
-
-}
