+++ src/main/Managers/Manager.java
@@ -1,7 +1,12 @@
+package Managers;
+// Developed by John Park and Rahul Jaideep
+
+import MatchingAlgoHelpers.MatchMaker;
+import MatchingAlgoHelpers.Scheduler;
 
 import java.util.ArrayList;
 
-abstract class Manager {
+public abstract class Manager {
     private MatchMaker matchMakerObject;
     private Scheduler schedulerObject;
 
@@ -10,35 +15,38 @@ abstract class Manager {
         this.schedulerObject = ss1;
     }
 
-    // Phase 1 - added this method so that it can be called in PriorityLocationManager subclass.
+    // Phase 1 - added this method so that it can be called in Manager.PriorityLocationManager subclass.
     // Needed to implement template design Pattern.
     public double getMatchableScoreLocation(Manageable user1, Manageable user2){
         return matchMakerObject.getScoreLocation(user1, user2);
     }
 
-    // Phase 1 - Added this method so that it can be called in PrioritySkillManager subclass.
+    // Phase 1 - Added this method so that it can be called in Manager.PrioritySkillManager subclass.
     // Needed to implement template design Pattern.
     public double getMatchableScoreSkill(Manageable user1, Manageable user2){
         return matchMakerObject.getScoreSkill(user1, user2);
     }
 
+    // This method is here for future use
     public ArrayList<String> getOverlapTimes(Manageable user1, Manageable user2){
         return schedulerObject.getCommonTimes(user1, user2);
     }
 
+    // This method is here for future use
     public boolean[] getSwipeStatusPair(Manageable user1, Manageable user2){
         boolean user1SwipeStatus = matchMakerObject.getSwipedStatus(user1, user2);
         boolean user2SwipeStatus = matchMakerObject.getSwipedStatus(user2, user1);
         return new boolean[]{user1SwipeStatus, user2SwipeStatus};
     }
 
+
     abstract double returnIndexedScorePriority(Manageable user1, Manageable user2);
 
     /**
      * Ranks the list of manageable objects based on their score with user1. Highest rank is index 0.
-     * @param user1 Manageable object (player)
-     * @param unranked_list List of Manageable objects (players) that are to be compared with user1 based on score.
-     * @return a ranked ArrayList of Manageable objects
+     * @param user1 Managers.Manageable object (player)
+     * @param unranked_list List of Managers.Manageable objects (players) that are to be compared with user1 based on score.
+     * @return a ranked ArrayList of Managers.Manageable objects
      */
     public ArrayList<Manageable> getRankedList(Manageable user1, ArrayList<Manageable> unranked_list){
         int listSize = unranked_list.size();
@@ -68,11 +76,14 @@ abstract class Manager {
         return returnList;
     }
 
+    /**
+     * This function uses the bubble sort algorithm to sort an array of Manageable items
+     * from lowest to highest indexedScores.
+     */
     public void bubbleSort(int listSize, double[] indexedScores, Manageable[] indexedManageables){
         for (int i = 0; i < listSize - 1; i++){
             for (int j = 0; j < listSize - i - 1; j++) {
                 if (indexedScores[j] > indexedScores[j + 1]) {
-                    // swap indexedScores[j+1] and indexedScores[j]
                     double temp1 = indexedScores[j];
                     indexedScores[j] = indexedScores[j + 1];
                     indexedScores[j + 1] = temp1;
