+++ src/main/MatchingAlgoHelpers/Schedulable.java
@@ -1,9 +1,9 @@
-/*
-    Developed by Aman Rana and Rahul Jaideep
- */
+package MatchingAlgoHelpers;
+// Developed by Aman Rana and Rahul Jaideep
+
 import java.util.ArrayList;
 
-interface Schedulable {
+public interface Schedulable {
     ArrayList<String> getSchedule();   //String list of times that the player is available in the week.
     String getUsername();
 }
