+++ src/main/UserData/BioRetrieval.java
@@ -1,0 +1,5 @@
+package UserData;
+
+public interface BioRetrieval {
+    String getBio();
+}
