+++ src/main/UserData/Game.java
@@ -1,8 +1,12 @@
-import java.util.Calendar;
-        import java.util.ArrayList;
-        import java.text.SimpleDateFormat;
+package UserData;
 
-public class Game implements BioRetrieval{
+import UserData.BioRetrieval;
+
+import java.util.Calendar;
+import java.util.ArrayList;
+import java.text.SimpleDateFormat;
+
+public class Game implements BioRetrieval {
 
     private Calendar date;
     private String court;
