+++ tests/GameTest.java
@@ -1,3 +1,4 @@
+import UserData.Game;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
