+++ tests/MatchMakerTest.java
@@ -1,4 +1,6 @@
-import org.junit.jupiter.api.BeforeEach;
+import MatchingAlgoHelpers.MatchMaker;
+import MatchingAlgoHelpers.Matchable;
+import UserData.Player;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
