+++ tests/PriorityLocationManagerTest.java
@@ -1,3 +1,8 @@
+import Managers.Manageable;
+import Managers.PriorityLocationManager;
+import MatchingAlgoHelpers.MatchMaker;
+import MatchingAlgoHelpers.Scheduler;
+import UserData.Player;
 import org.junit.jupiter.api.Test;
 
 import java.util.ArrayList;
